Method for getting possible segmentations is initiliazed.

+++ src/main/java/core/NestedSegmenter.java
@@ -15,28 +15,28 @@ import java.util.TreeMap;
  */
 public class NestedSegmenter {
 
-    private Map<String, Integer> stems = new TreeMap<String, Integer>();
-    private Map<String, Integer> affixes = new TreeMap<String, Integer>();
-    private Map<String, Integer> results = new TreeMap<String, Integer>();
-    private Map<String, Integer> notFound = new TreeMap<String, Integer>();
+    private Map<String, Double> stems = new TreeMap<String, Double>();
+    private Map<String, Double> affixes = new TreeMap<String, Double>();
+    private Map<String, Double> results = new TreeMap<String, Double>();
+    private Map<String, Double> notFound = new TreeMap<String, Double>();
 
     private String fileSegmentationInput;
 
     private WordVectors vectors;
 
-    public Map<String, Integer> getStems() {
+    public Map<String, Double> getStems() {
         return stems;
     }
 
-    public Map<String, Integer> getAffixes() {
+    public Map<String, Double> getAffixes() {
         return affixes;
     }
 
-    public Map<String, Integer> getResults() {
+    public Map<String, Double> getResults() {
         return results;
     }
 
-    public Map<String, Integer> getNotFound() {
+    public Map<String, Double> getNotFound() {
         return notFound;
     }
 
@@ -49,7 +49,7 @@ public class NestedSegmenter {
         }
     }
 
-    private void doNested(String word, int frequency, double treshold) {
+    private void doNested(String word, double frequency, double treshold) {
 
         Stack<String> localSuffixes = new Stack<String>();
         String stem = word;
@@ -121,7 +121,7 @@ public class NestedSegmenter {
                 String space = " ";
                 StringTokenizer st = new StringTokenizer(line, space);
 
-                int freq = Integer.parseInt(st.nextToken());
+                double freq = Double.parseDouble(st.nextToken());
                 String word = st.nextToken();
 
                 doNested(word, freq, 0.25);
@@ -136,34 +136,34 @@ public class NestedSegmenter {
         NestedSegmenter ns = new NestedSegmenter(args[0], args[1]);
         ns.findSegmentsAndAffixes();
 
-        Map<String, Integer> s = ns.getStems();
-        Map<String, Integer> a = ns.getAffixes();
-        Map<String, Integer> r = ns.getResults();
-        Map<String, Integer> n = ns.getNotFound();
+        Map<String, Double> s = ns.getStems();
+        Map<String, Double> a = ns.getAffixes();
+        Map<String, Double> r = ns.getResults();
+        Map<String, Double> n = ns.getNotFound();
 
         PrintWriter writer_seg = new PrintWriter("outputs/stems", "UTF-8");
         PrintWriter writer_af = new PrintWriter("outputs/affixes", "UTF-8");
         PrintWriter writer_res = new PrintWriter("outputs/results", "UTF-8");
         PrintWriter writer_noF = new PrintWriter("outputs/absent", "UTF-8");
 
-        for (Map.Entry<String, Integer> entry : s.entrySet()) {
+        for (Map.Entry<String, Double> entry : s.entrySet()) {
             String line = entry.getValue() + " " + entry.getKey();
             writer_seg.println(line);
         }
         writer_seg.close();
 
-        for (Map.Entry<String, Integer> entry : a.entrySet()) {
+        for (Map.Entry<String, Double> entry : a.entrySet()) {
             String line = entry.getValue() + " " + entry.getKey();
             writer_af.println(line);
         }
         writer_af.close();
 
-        for (Map.Entry<String, Integer> entry : r.entrySet()) {
+        for (Map.Entry<String, Double> entry : r.entrySet()) {
             String line = entry.getValue() + " " + entry.getKey();
             writer_res.println(line);
         }
 
-        for (Map.Entry<String, Integer> entry : n.entrySet()) {
+        for (Map.Entry<String, Double> entry : n.entrySet()) {
             String line = entry.getValue() + " " + entry.getKey();
             writer_noF.println(line);
         }
